Version 0.2
ant.vim will now add a 'Target' menu when the user manually sets the location of
the build file with ',s' or by selecting the build option.  So, for example, selecting
ANT->Target->clean from the menu will execute 'ant clean.'  I'm not all that confident
in my regex skills, so I can't guarantee that it will work in all (or even most!) circumstances.
Any help or bug reports will be appreciated.