Version 0.4
Added ',g' keyboard shortcut.  When in the ant error buffer, this
will attempt to parse out the file in error and open it in a new
buffer with the cursor at the line flagged in error.  The user must
execute the shortcut at the line that identifies the file in error.